Move pipeResponse into a more appropriately named file.
